[TML+doc] Restructure mortality modules/concepts
