+++ Documentation/ProjectFile/MangaProject/tree_dynamics/tree_growth_and_death/Mortality/NoGrowth/c_NoGrowth.md
@@ -1,0 +1,7 @@
+
+`NoGrowth` is a mortality concept.
+
+In `NoGrowth`, a tree dies if the growth variable is equal or below 0, i.e. no growth happens. 
+
+This concept is based on mechanistic causes, e.g. resource deficit, and no tree dies of random reasons.
+
