+++ Documentation/ProjectFile/MangaProject/tree_dynamics/tree_growth_and_death/Mortality/Random/c_Random.md
@@ -1,0 +1,10 @@
+
+`Random` is a mortality concept.
+
+In `Random`, a tree dies with a defined yearly mortality probability.
+If a random number times the number of time steps per year is below this defined probability the tree dies.
+The default probability is 0.2 %. Berger & Hildenbrandt (2000) suggested 0.2 and 0.16 % for Rhizophora mangle and Avicennia germinans, respectively.  
+Use `<probability> 0.0016 </probability>` to modify concept parameters.
+
+In this concept, a tree does not die of mechanistic reasons. It is useful to combine `Random` with `NoGrowth` or `Memory`.  
+
