--- Documentation/project_dox/MangaProject/tree_dynamics/tree_growth_and_death/Mortality/c_Mortality.dox
@@ -1,50 +1,0 @@
-/*! \page project_dox__MangaProject__tree_dynamics__tree_growth_and_death__Mortality__Mortality &emsp;[case]Mortality
-
-
-**Mortality** concepts define why trees die.  
-The following concepts are available for `SimpleBettina` and `NetworkBettina`
-
-- `NoGrowth`: trees die if growth is <= 0
-- `Random`: trees die randomly with a defined probability
-- `Memory`: trees die if relative growth falls below a given threshold
-
-Mortality concepts are defined using the tag `<mortality>`. 
-The default concept is `NoGrowth` (no specification in the input xml file required).
-Concepts can be combined by listing them in `<mortality>`, separated by a whitespace, e.g. `<mortality> random memory </mortality>`.
-  
-
-### Concept `NoGrowth`
-
-In `NoGrowth`, a tree dies if the growth variable is equal or below 0, i.e. no growth happens. 
-
-This concept is based on mechanistic causes, e.g. resource deficit, and no tree dies of random reasons.
-
-
-### Concept `Random`
-
-In `Random`, a tree dies with a defined yearly mortality probability.
-If a random number times the number of time steps per year is below this defined probability the tree dies.
-The default probability is 0.2 %. Berger & Hildenbrandt (2000) suggested 0.2 and 0.16 % for Rhizophora mangle and Avicennia germinans, respectively.  
-Use `<probability> 0.002 </probability>` to modify concept parameters.
-
-In this concept, a tree does not die of mechanistic reasons. It is useful to combine `Random` with `NoGrowth` or `Memory`.  
-
-
-### Concept `Memory`
-
-In `Memory`, a tree remembers its relative growth of the preceding N years (in m³ per memory period). 
-Relative growth is defined as the fraction of growth in tree biomass (in m³ per m³).
-If the relative growth falls below a certain threshold, a tree dies.  
-
-The default threshold is 0.5 % (i.e. `<threshold> 0.005 </threshold>`).  
-The default memory period is 1 year (i.e. `<period> 31557600 </period>`).
-
-This concept is based on mechanistic causes, e.g. slowing down of growth due external perturbations. 
-A tree can survive a period of resource scarcity but if conditions do not improve, the tree dies.  
-No tree dies of random reasons.
-# Required inputs: 
- 
-- \subpage project_dox__MangaProject__tree_dynamics__tree_growth_and_death__Mortality__period
-- \subpage project_dox__MangaProject__tree_dynamics__tree_growth_and_death__Mortality__probability
-- \subpage project_dox__MangaProject__tree_dynamics__tree_growth_and_death__Mortality__threshold
-*/
