--- Documentation/project_dox/MangaProject/tree_dynamics/tree_growth_and_death/Mortality/s_period.dox
@@ -1,5 +1,0 @@
-/*! \page project_dox__MangaProject__tree_dynamics__tree_growth_and_death__Mortality__period &emsp;[spec]period
-
-`period` defines the duration a tree remembers it growth process in seconds, described by the biomass increment per time step.   
-The default is 157,788,000 seconds (= 5 years).
-*/
