+++ TreeModelLib/GrowthAndDeathDynamics/Mortality/NoGrowth/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+@date: 2021-Today
+@author: marie-christin.wimmler@tu-dresden.de
+"""
+
+from .NoGrowth import NoGrowth
